+++ core_utils.py
